pyproject.toml
[project]
name = "sentinelops"
version = "0.1.0"
description = "Incident Response Triage Agent — monitors Elasticsearch, detects anomalies, and creates prioritized incidents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch[async]>=8.12,<9",
    "anthropic>=0.40,<1",
    "slack-sdk>=3.27,<4",
    "pdpyras>=5.2,<6",
    "pyyaml>=6.0,<7",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "structlog>=24.1,<25",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-mock>=3.12,<4",
]

[project.scripts]
sentinelops = "sentinelops.main:cli_entry"

[tool.setuptools.packages.find]
include = ["sentinelops*"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
